refactor(server): isolate addmissingates pandas implem
