+++ server/weaverbird/backends/pandas_executor/steps/__init__.py
@@ -1,8 +1,8 @@
 from typing import Dict
 
 from ..types import StepExecutor
+from .addmissingdates import execute_addmissingdates
 
-# from .addmissingdates import execute_addmissingdates
 # from .aggregate import execute_aggregate
 # from .append import execute as execute_append
 # from .argmax import execute as execute_argmax
